requirement.txt
discord-rich-presence==1.1.0
pypresence==4.3.0
python-dotenv==1.1.0
python-mpd2==3.1.1
